requirements.txt
SQLAlchemy==2.0.30
python-dotenv==1.0.0
aiogram==3.0.0b7
requests==2.32.3
yookassa==3.1.0